Move common styles to utilities
