+++ packages/charts/react-charting/src/components/AreaChart/AreaChart.styles.ts
@@ -1,5 +1,5 @@
 import { IAreaChartStyleProps, IAreaChartStyles } from './AreaChart.types';
-import { getTooltipStyles } from '../../CartesianChart';
+import { getTooltipStyles } from '../../utilities';
 
 export const getStyles = (props: IAreaChartStyleProps): IAreaChartStyles => {
   return {
