+++ packages/charts/react-charting/src/components/CommonComponents/index.ts
@@ -1,3 +1,2 @@
 export * from './CartesianChart';
 export * from './CartesianChart.types';
-export * from './Common.styles';
