+++ packages/charts/react-charting/src/components/LineChart/LineChart.styles.ts
@@ -1,6 +1,6 @@
 import { HighContrastSelector } from '@fluentui/react/lib/Styling';
 import { ILineChartStyleProps, ILineChartStyles } from './LineChart.types';
-import { getTooltipStyles } from '../../CartesianChart';
+import { getTooltipStyles } from '../../utilities';
 
 export const getStyles = (props: ILineChartStyleProps): ILineChartStyles => {
   return {
