+++ packages/charts/react-charting/src/utilities/index.ts
@@ -1,4 +1,5 @@
 export * from './ChartHoverCard/index';
+export * from './Common.styles';
 export * from './utilities';
 export * from './colors';
 export * from './gradients';
